fixed mismatch type

+++ RR-Common/src/main/java/com/scor/rr/domain/DefaultAdjustmentRegionPerilEntityPK.java
@@ -25,7 +25,7 @@ public class DefaultAdjustmentRegionPerilEntityPK implements Serializable {
         return fkRegionPerilId;
     }
 
-    public void setFkRegionPerilId(int fkRegionPerilId) {
+    public void setFkRegionPerilId(long fkRegionPerilId) {
         this.fkRegionPerilId = fkRegionPerilId;
     }
